Add files via upload
This is a simple Javascript game called Americaw coded as part of my Multimedia Programming module in second year. In it, you play as the famous Eagle that attacked Donald Trump in the photoshoot earlier this year. The aim of the game is to "Trump" the Trumps that show up on screen to save all 50 states of America, while avoid Bernie Sanders. 
Simply use the arrow keys to move to target the faces. When you hit one, a sound bite will play (this doesn't work on firefox, it does work on chrome).